Refactor stream module into more files in its own folder
