+++ src/pileup_augmenter.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept>
 #include "json2pb.h"
 #include "pileup_augmenter.hpp"
-#include "stream.hpp"
+#include "stream/stream.hpp"
 
 using namespace std;
