+++ src/readfilter.cpp
@@ -1,6 +1,6 @@
 #include "readfilter.hpp"
 #include "IntervalTree.h"
-#include "stream.hpp"
+#include "stream/stream.hpp"
 
 #include <fstream>
 #include <sstream>
