+++ src/snarls.hpp
@@ -14,7 +14,8 @@
 #include <unordered_set>
 #include <fstream>
 #include <deque>
-#include "stream.hpp"
+#include "stream/protobuf_emitter.hpp"
+#include "stream/protobuf_iterator.hpp"
 #include "vg.hpp"
 #include "handle.hpp"
 #include "vg.pb.h"
