+++ src/stream/hfile_cppstream.cpp
