+++ src/stream/hfile_cppstream.hpp
