+++ src/stream/stream.cpp
