+++ src/subcommand/align_main.cpp
@@ -15,7 +15,7 @@
 
 #include "../utility.hpp"
 #include "../vg.hpp"
-#include "../stream.hpp"
+#include "../stream/stream.hpp"
 
 using namespace std;
 using namespace vg;
