+++ src/subcommand/genotype_main.cpp
@@ -2,11 +2,11 @@
 #include <getopt.h>
 #include "subcommand.hpp"
 #include "index.hpp"
-#include "stream.hpp"
+#include "stream/stream.hpp"
 #include "genotyper.hpp"
 #include "genotypekit.hpp"
 #include "variant_recall.hpp"
-#include "stream.hpp"
+#include "stream/stream.hpp"
 /**
 * GAM sort main
 */
