+++ src/subcommand/locify_main.cpp
@@ -15,7 +15,7 @@
 #include "../vg.hpp"
 #include "../index.hpp"
 #include "../convert.hpp"
-#include "../stream.hpp"
+#include "../stream/stream.hpp"
 
 using namespace std;
 using namespace vg;
