+++ src/subcommand/mod_main.cpp
@@ -12,7 +12,7 @@
 
 #include "../vg.hpp"
 #include "../cactus.hpp"
-#include "../stream.hpp"
+#include "../stream/stream.hpp"
 #include "../utility.hpp"
 #include "../algorithms/topological_sort.hpp"
 #include "../algorithms/remove_high_degree.hpp"
