+++ src/subcommand/recalibrate_main.cpp
@@ -11,7 +11,7 @@
 
 #include "../alignment.hpp"
 #include "../vg.hpp"
-#include "../stream.hpp"
+#include "../stream/stream.hpp"
 
 #include <vowpalwabbit/vw.h>
