+++ src/subcommand/vectorize_main.cpp
@@ -16,7 +16,7 @@
 #include "../vg.hpp"
 #include "../vectorizer.hpp"
 #include "../mapper.hpp"
-#include "../stream.hpp"
+#include "../stream/stream.hpp"
 
 using namespace std;
 using namespace vg;
