+++ src/unittest/blocked_gzip_output_stream.cpp
@@ -2,8 +2,8 @@
 ///  
 /// Unit tests for BlockedGzipOutputStream 
 
-#include "../blocked_gzip_output_stream.hpp"
-#include "../hfile_cppstream.hpp"
+#include "../stream/blocked_gzip_output_stream.hpp"
+#include "../stream/hfile_cppstream.hpp"
 #include "catch.hpp"
 
 #include <htslib/hfile.h>
